+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Bug fixes:
 * Fixed `TracePoint` thread-safety by storing the state on the Ruby `Thread` (like MRI) instead of inside the `TracePoint` instance.
 * Make `require 'rubygems/package'` succeed and define `Gem::Deprecate` correctly (#2014).
 * Fix `MBCLEN_CHARFOUND_P` error.
+* Fix `rb_enc_str_new` when `NULL` encoding is given with a constant string.
 
 Compatibility: